requirements.txt
exifread
Pillow
piexif
opencv-python
json
